refactor: code refactoring
Naming change and package structure change
Code deduplication
