+++ src/main/java/com/fastcampus/investment/service/ProductService.java
@@ -2,7 +2,7 @@ package com.fastcampus.investment.service;
 
 import com.fastcampus.investment.domain.Investment;
 import com.fastcampus.investment.domain.Product;
-import com.fastcampus.investment.dto.ProductResponse;
+import com.fastcampus.investment.dto.response.ProductResponse;
 import com.fastcampus.investment.exception.APIException;
 import com.fastcampus.investment.repository.InvestmentRepository;
 import com.fastcampus.investment.repository.ProductRepository;
